Issue#2 Adjust Template Content (#5)
* Update branding from eCommerce to Supply Shop

Changed the webpage title, navigation bar brand name, and footer text to reflect the new branding as "Supply Shop".

* Update footer to display current year dynamically

The footer in _Layout.cshtml has been modified to replace the hardcoded year (2025) with a dynamic display of the current year using "@DateTime.Today.Year". This change ensures that the year is always up-to-date.

* Update welcome message styling and add construction alert

Modified `Index.cshtml` to change the `<h1>` class from `display-4` to `fs-1 fw-normal`. Added an alert to inform users that the website is currently under construction.